+++ modules/alloy/src/main/ts/ephox/alloy/api/behaviour/Positioning.ts
@@ -1,7 +1,7 @@
 import * as ActivePosition from '../../behaviour/positioning/ActivePosition';
 import * as PositionApis from '../../behaviour/positioning/PositionApis';
 import { PositioningBehaviour } from '../../behaviour/positioning/PositioningTypes';
-import PositionSchema from '../../behaviour/positioning/PositionSchema';
+import { PositionSchema } from '../../behaviour/positioning/PositionSchema';
 import * as Behaviour from './Behaviour';
 
 const Positioning: PositioningBehaviour = Behaviour.create({