+++ modules/katamari/CHANGELOG.md
@@ -20,6 +20,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - The `Result` type is now covariant with respect to its type argument.
 - All singletons now have a `get` function that returns an `Optional` value.
 
+### Fixed
+- `Throttler.adaptable` could not re-throttle from within the callback function.
+
 ## 7.2.0 - 2021-05-06
 
 ### Added